Added logic to restrict player movements.
Created a new enum class direction. Added new method setWallCoordinates in MapLoader Class which creates a new list containing all the coordinates of all the walls.
Added new method isWall which checks if there is a wall near the given coordinates.
Added new mehtod canPlayerMove which checks if the player can move in the given direction.