Updated GameScreen & MapLoader
Added method setCoordinateList to MapLoader which add the coordinates to coordinateLists.
Added method isObstacle which checks if the player is on the obstacle.
Added method handelObstacle which reduces hearts if the player touches obstacle and after every three seconds of player staying on the obstacle.
Added method handelWin & handelLoose which on Win or Loose changes the screen to MenuScreen.